requirement.txt
asgiref==3.8.1
Django==4.2.9
django-debug-toolbar==4.4
numpy==2.0.0
pillow==11.0.0
python-dateutil==2.8.2
pytz==2024.2
sqlparse==0.5.2
